DDPG

Documentation for DDPG.

This implementation of DDPG is based on the paper Continuous control with deep reinforcement learning by Lillicrap et al. (2015). It solves the following environments:

Pendulum

Lunar Lander

Installation

In the julia REPL, run

using Pkg
Pkg.add(url="https://github.com/SvenDuve/DDPG.jl")

This package is not for general use. To be fully functional, it requires the following packages:

RLTypes

Conda

PyCall

Gymnasium, this is a slightly adapted version. The original package can be found here

Run the following:

using Pkg
Pkg.add(url="https://github.com/SvenDuve/RLTypes.jl")
Pkg.add("Conda")
Pkg.add("PyCall")

Then again within julia,

using Conda
using PyCall
Conda.pip("install", "git+https://github.com/SvenDuve/Gymnasium")

Bring package into scope with

using DDPG

Example usage

Once the above is installed, the following code can be used to train an agent:

using DDPG
using RLTypes
agent(LunarLander(), AgentParameter(training_episodes=100, batch_size=128))

This function returns a data structure containing a trained agent.

Function Reference

DDPG.action

DDPG.agent

DDPG.setActor

DDPG.setCritic

DDPG.soft_update!

DDPG.train_step!

DDPG.action — Method

action(model, state, train, ep, hp)

Returns an action based on the current state, noise and the model.

DDPG.agent — Method

agent(environment::ContinuousEnvironment, agentParams::AgentParameter)

Main DDPG Algorithm to train an agent.

DDPG.setActor — Method

setActor(state_size, action_size)

Set an actor model for the DDPG agent.

DDPG.setCritic — Method

setCritic(state_size, action_size)

Set a critic model for the DDPG agent.

DDPG.soft_update! — Method

soft_update!(target_model, main_model, τ)

Soft update of the target model.

DDPG.train_step! — Method

train_step!(S, A, R, S´, T, μθ, μθ´, Qϕ, Qϕ´, opt_critic, opt_actor, ap::Parameter)

Train the DDPG agent with gradient descent.